lots of progress, it's working! see description
translate/scaling of baked attractors now work. scaling kind of works. need to make sure the field size and the attractor size are independent of one another. key insight was scaling up the size of the visualizer to be 256x256 since that's the "space" the baked map is generated in. also realized i needed to adjust the mass of an object with the square of its scale!

current weirdness: objects do something weird when they fall through the centre of a baked attractor's map, they accelerate away. even in a circular baked map where there should be basically zero force. somethings up.