No real progress, just experiementation with gravity simulation
Spent ages watching the way attractors oscillate up and down inside shell attractors, it's a good litmus test for the simulation: if the object always returns to where it started, things are good. As well as the previous note about having "interpolation" set to "none" on rigidbodies, it seems the "async" mode of the gravity simulation also breaks it.

The other concern is that the way the character controller adds gravity doesn't seem to match how the attractors do it, but the tutorial I used insists on setting velocity instead of adding it or something, so I have to figure out how to change it to match so the attractor can just add gravity without the character controller having to worry about it.