Jumping is unjanked
I have to be honest, I'm not entirely sure why this is fixed now. I took a closer look at how the "non rigidbody attractor" was working and applied it to the character controller, but the end result looks pretty much the same as the catlike coding tutorial, but it works. Hmm...